Refactor exercise selection feature: implement new ExerciseSelectionView, replace bottom sheet with navigation, and enhance workout bucket management
