+++ lib/core/widgets/custom_divider.dart
@@ -1,0 +1,25 @@
+import 'package:flutter/material.dart';
+import 'package:superset/core/constants/app_padings.dart';
+
+final class CustomDivider extends StatelessWidget {
+  const CustomDivider({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      height: 1,
+      margin: const AppPaddings.normal(),
+      decoration: BoxDecoration(
+        gradient: LinearGradient(
+          colors: [
+            Colors.transparent,
+            Colors.grey.withOpacity(0.3),
+            Colors.transparent,
+          ],
+        ),
+      ),
+    );
+  }
+}
