+++ lib/features/workout/cubit/workout_state.dart
@@ -8,6 +8,7 @@ final class WorkoutState extends Equatable {
     this.isLoading = false,
     this.exercises,
     this.workout,
+    this.workoutBucket,
   });
 
   //ilk açıldıgında selectedDate'i bugünün tarihi olarak ayarlıyoruz
@@ -21,24 +22,34 @@ final class WorkoutState extends Equatable {
   /// This is used for all exercises for the app.
   final List<Exercise>? exercises;
 
+  final Workout? workoutBucket;
+
   ///tarihi seçtiğinde firebase'e istek atılacak eğer o tarihte bir workout varsa o gelecek
   /// eğer yoksa add exercise butonuna basınca yeni bir workout oluşturulacak
   final Workout? workout;
 
   @override
-  List<Object?> get props => [selectedDate, exercises, isLoading, workout];
+  List<Object?> get props => [
+    selectedDate,
+    exercises,
+    isLoading,
+    workout,
+    workoutBucket,
+  ];
 
   WorkoutState copyWith({
     DateTime? selectedDate,
     List<Exercise>? exercises,
     bool? isLoading,
     Workout? workout,
+    Workout? workoutBucket,
   }) {
     return WorkoutState(
       selectedDate: selectedDate ?? this.selectedDate,
       exercises: exercises ?? this.exercises,
       isLoading: isLoading ?? this.isLoading,
       workout: workout ?? this.workout,
+      workoutBucket: workoutBucket ?? this.workoutBucket,
     );
   }
 }
